+++ src/data/JS/agda-rts.js
@@ -185,7 +185,6 @@ exports.primFloatIsSafeInteger = function(x) {
 
 
 // These WORD64 values were obtained via `castDoubleToWord64` in Haskell:
-const WORD64_NAN      = 18444492273895866368n;
 const WORD64_POS_INF  = 9218868437227405312n;
 const WORD64_NEG_INF  = 18442240474082181120n;
 const WORD64_POS_ZERO = 0n;
@@ -193,7 +192,7 @@ const WORD64_NEG_ZERO = 9223372036854775808n;
 
 exports.primFloatToWord64 = function(x) {
     if (exports.primFloatIsNaN(x)) {
-        return WORD64_NAN;
+        return null;
     }
     else if (x < 0.0 && exports.primFloatIsInfinite(x)) {
         return WORD64_NEG_INF;